feat: capability to load models in background (android) w.i.p
