+++ android/src/main/java/com/visionsdk/VisionSdkPackage.kt
@@ -8,7 +8,7 @@ import com.facebook.react.uimanager.ViewManager
 
 class VisionSdkPackage : ReactPackage {
   override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
-    return emptyList()
+    return listOf(VisionSdkModule(reactContext))
   }
 
   override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {
