+++ example/package.json
@@ -10,18 +10,24 @@
   },
   "dependencies": {
     "@react-native/metro-config": "^0.74.87",
+    "@react-navigation/native": "^7.0.14",
+    "@react-navigation/stack": "^7.1.1",
     "react": "18.3.1",
     "react-native": "0.76.6",
+    "react-native-gesture-handler": "^2.24.0",
     "react-native-permissions": "^4.1.5",
     "react-native-progress": "^5.0.1",
+    "react-native-reanimated": "^3.16.7",
+    "react-native-safe-area-context": "^5.2.0",
+    "react-native-screens": "^4.9.0",
     "react-native-svg": "^15.11.1",
-    "react-native-vector-icons": "^10.1.0"
+    "react-native-vector-icons": "^10.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
     "@babel/runtime": "^7.25.0",
-    "@react-native-community/cli-platform-ios": "^15.1.3",
     "@react-native-community/cli": "latest",
+    "@react-native-community/cli-platform-ios": "^15.1.3",
     "@react-native/babel-preset": "0.76.6",
     "@types/react-native-vector-icons": "^6.4.18",
     "babel-plugin-module-resolver": "^5.0.2",
